Create Security Audit
This code will allow you to input a target to scan, and will then scan a list of common ports to see if they are open. If any of the ports are open, it will print out which ones are open. If you enter 22 (SSH), it will ask if you would like to scan more ports, and if you enter 80 (HTTP), it will exit.