Create DEVIL EYE 3.0
This tool is called "Devil Eye" and it is a WiFi monitoring and hacking tool. It is designed to scan the network for devices and display a list of all connected devices. The user can then select a device from the list by entering its ID number and choose to "kick" it off the network by sending a deauthentication packet to the device's MAC address. This tool can be used to detect and remove unauthorized devices from a network.