Update DEVIL EYE 4.0
This Python script is a WiFi monitoring tool called "Devil Eye" that allows users to scan a network for devices and kick them off the network. The script uses Scapy to send deauthentication packets to the targeted device, forcing it to disconnect from the network. The user can select the device they want to kick by entering its ID number from the list of devices detected by the script. The script also includes command-line arguments for specifying the network interface to use for packet injection and the IP range to scan for devices. The script logs kicked devices in a file named "devileye.log".